icap_gui/requirements.txt
PySide6~=6.4.2
pyserial~=3.5